+++ managed/yba-installer/cmd/install.go
@@ -65,5 +65,9 @@ func init() {
 	installCmd.Flags().StringSliceVarP(&skippedPreflightChecks, "skip_preflight", "s",
 		[]string{}, "Preflight checks to skip")
 	installCmd.Flags().StringVarP(&licensePath, "license-path", "l", "", "path to license file")
-	rootCmd.AddCommand(installCmd)
+
+	// Install must be run from directory of yba version
+	if !common.RunFromInstalled() {
+		rootCmd.AddCommand(installCmd)
+	}
 }