+++ src/yb/yql/pggate/pggate.cc
@@ -564,6 +564,25 @@ Status PgApiImpl::UpdateSequenceTuple(int64_t db_oid,
   return Status::OK();
 }
 
+Status PgApiImpl::FetchSequenceTuple(int64_t db_oid,
+                                     int64_t seq_oid,
+                                     uint64_t ysql_catalog_version,
+                                     bool is_db_catalog_version_mode,
+                                     uint32_t fetch_count,
+                                     int64_t inc_by,
+                                     int64_t min_value,
+                                     int64_t max_value,
+                                     bool cycle,
+                                     int64_t *first_value,
+                                     int64_t *last_value) {
+  auto res = VERIFY_RESULT(pg_session_->FetchSequenceTuple(
+      db_oid, seq_oid, ysql_catalog_version, is_db_catalog_version_mode, fetch_count, inc_by,
+      min_value, max_value, cycle));
+  *first_value = res.first;
+  *last_value = res.second;
+  return Status::OK();
+}
+
 Status PgApiImpl::ReadSequenceTuple(int64_t db_oid,
                                     int64_t seq_oid,
                                     uint64_t ysql_catalog_version,
@@ -1833,6 +1852,11 @@ void PgApiImpl::SetTimeout(int timeout_ms) {
 
 Result<client::TabletServersInfo> PgApiImpl::ListTabletServers() {
   return pg_session_->ListTabletServers();
+}
+
+Status PgApiImpl::GetIndexBackfillProgress(std::vector<PgObjectId> oids,
+                                           uint64_t** backfill_statuses) {
+  return pg_session_->GetIndexBackfillProgress(oids, backfill_statuses);
 }
 
 Status PgApiImpl::ValidatePlacement(const char *placement_info) {